feat: rename banner component
