+++ src/ui/Banner/Banner.module.scss
