+++ src/ui/Banner/index.ts
@@ -1,0 +1,1 @@
+export { default } from './Banner';
