+++ src/ui/index.ts
@@ -1,6 +1,6 @@
 export { default as Adornment } from './Adornment';
 
-export { default as Alert } from './Alert';
+export { default as Banner } from './Banner';
 
 export { default as Avatar } from './Avatar';
 export type { AvatarProps } from './Avatar';
